Move the address lookup to the test package for the time being
