+++ conf/prod.routes
@@ -1,6 +1,6 @@
 # Add all the application routes to the app.routes file
 
-->         /fset-fast-track                app.Routes
+->         /fset-fast-track               app.Routes
 ->         /                              health.Routes
 
 GET        /admin/metrics                 com.kenshoo.play.metrics.MetricsController.metrics
