+++ conf/testOnlyDoNotUseInAppConf.routes
@@ -12,4 +12,4 @@
 # Add all the application routes to the prod.routes file
 ->         /                          prod.Routes
 
-GET        /address/search/:postcode             controllers.AddressLookupController.addressLookup(postcode: String)
+GET     /fset-fast-track/address-search/:postcode                              controllers.test.AddressLookupController.addressLookup(postcode: String)
